refactor(operator): extract SecurityTab from SecuritySection
Moves SecuritySection's full UI into SecurityTab; section is always
defaultExpanded=true on the dedicated sub-route. SecuritySection.tsx
and its test are deleted. Operator.tsx has SecuritySection references
removed.
